refactor(Icon20px): move font icons into internal/icons directory
