+++ packages/retail-ui/components/TopBar/Item.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import CapIcon, { IconProps } from '../Icon/20px';
+import CapIcon, { IconProps } from '../internal/icons/20px';
 import styles from './TopBar.module.less';
 import { createPropsGetter } from '../internal/createPropsGetter';
 import { cx } from '../../lib/theming/Emotion';
