+++ packages/retail-ui/components/TopBar/TopBarDropdown.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { isKeyArrowVertical, isKeyEnter, isKeySpace, someKeys } from '../../lib/events/keyboard/identifiers';
 import { Nullable } from '../../typings/utility-types';
 import ButtonItem from './ButtonItem';
-import { IconProps } from '../Icon/20px';
+import { IconProps } from '../internal/icons/20px';
 import DropdownMenu from '../DropdownMenu';
 import { PopupMenuCaptionProps } from '../internal/PopupMenu/PopupMenu';
