+++ packages/retail-ui/components/internal/icons/20px/Icon.js.flow
