+++ packages/retail-ui/components/internal/icons/20px/Icon.module.less
