+++ packages/retail-ui/components/internal/icons/20px/Kontur-Iconic-eco-20px.woff2
