+++ packages/retail-ui/components/internal/icons/20px/index.tsx
